# openapi.yaml
openapi: 3.0.3
info:
  title: FinApp API
  description: API for financial application
  version: 1.3.0
  contact:
    name: API Support
    email: support@finapp.com
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
  - url: https://api.finapp.com/api/v1
    description: Production server
paths:
...
  /analytics:
    get:
      operationId: getAnalytics
      tags: [Analytics]
      summary: Get analytics data
      description: |
        Возвращает аналитическую информацию о доходах, расходах, балансе и последних транзакциях за указанный период.
        По умолчанию, период составляет последний год.
      parameters:
        - name: startDate
          in: query
          description: Начальная дата периода в формате YYYY-MM-DD
          required: false
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          description: Конечная дата периода в формате YYYY-MM-DD
          required: false
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Analytics data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyticsResponse'
        '400':
          description: Invalid parameters
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /analytics/categories:
    get:
      operationId: getCategoriesAnalytics
      tags: [Analytics]
      summary: Get categories analytics
      description: |
        Возвращает аналитику по категориям - доходы и расходы сгруппированные по категориям за период.
      parameters:
        - name: startDate
          in: query
          description: Начальная дата периода в формате YYYY-MM-DD
          required: false
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          description: Конечная дата периода в формате YYYY-MM-DD
          required: false
          schema:
            type: string
            format: date
        - name: type
          in: query
          description: Тип транзакций для анализа
          required: false
          schema:
            $ref: '#/components/schemas/TransactionType'
      responses:
        '200':
          description: Categories analytics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryAnalyticsResponse'
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /analytics/summary:
    get:
      operationId: getPeriodSummary
      tags: [Analytics]
      summary: Get period summary
      description: |
        Возвращает сводную аналитику за указанный период с группировкой по месяцам.
      parameters:
        - name: startDate
          in: query
          description: Начальная дата периода в формате YYYY-MM-DD
          required: false
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          description: Конечная дата периода в формате YYYY-MM-DD
          required: false
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Period summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PeriodSummaryResponse'
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /analytics/daily:
    get:
      operationId: getDailyTransactions
      tags: [Analytics]
      summary: Get daily transactions
      description: |
        Возвращает ежедневную статистику транзакций за период с группировкой по дням.
      parameters:
        - name: startDate
          in: query
          description: Начальная дата периода в формате YYYY-MM-DD
          required: false
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          description: Конечная дата периода в формате YYYY-MM-DD
          required: false
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Daily transactions statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DailyTransactionsResponse'
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'

  # Transaction endpoints
...
components:
  schemas:
...
    TransactionDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: "ID транзакции"
        amount:
          type: string
          pattern: '^\d+\.\d{2}$'
          description: "Сумма транзакции в формате decimal string (копейки)"
          example: "250.50"
        currency:
          type: string
          pattern: '^[A-Z]{3}$'
          description: "Код валюты ISO-4217"
          example: "RUB"
          default: "RUB"
        date:
          type: string
          format: date
          description: "Дата транзакции"
          example: "2025-09-15"
        categoryId:
          type: string
          format: uuid
          description: "ID категории"
        name:
          type: string
          description: "Название транзакции"
          example: "Обед в кафе"
        type:
          $ref: '#/components/schemas/TransactionType'
        inputType:
          $ref: '#/components/schemas/InputType'
        accountId:
          type: string
          format: uuid
          description: "ID счета"
        excluded:
          type: boolean
          description: "Исключить из расчетов"
          default: false
...
    TransactionType:
      type: string
      enum: [INCOME, EXPENSE, TRANSFER]
      description: "Тип транзакции"
    InputType:
      type: string
      enum: [MANUAL, IMPORT, RECEIPT_SCAN]
      description: "Тип ввода транзакции"
...
    ProblemDetails:
      type: object
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
          description: "URI идентифицирующий тип проблемы"
          example: "https://api.finapp.com/problems/validation-error"
        title:
          type: string
          description: "Краткое описание типа проблемы"
          example: "Validation Error"
        status:
          type: integer
          description: "HTTP статус код"
          example: 400
        detail:
          type: string
          description: "Детальное описание проблемы"
          example: "Поле 'amount' должно быть положительным числом"
        instance:
          type: string
          format: uri
          description: "URI идентифицирующий конкретный случай проблемы"
          example: "/api/v1/transactions/123"
        errors:
          type: array
          description: "Детали ошибок валидации"
          items:
            type: object
            properties:
              field:
                type: string
                description: "Поле с ошибкой"
                example: "amount"
              message:
                type: string
                description: "Сообщение об ошибке"
                example: "должно быть положительным числом"
              code:
                type: string
                description: "Код ошибки"
                example: "POSITIVE_NUMBER_REQUIRED"

    # User DTOs
...
    AnalyticsResponse:
      type: object
      properties:
        metrics:
          type: object
          properties:
            income:
              type: string
              pattern: '^\d+\.\d{2}$'
              description: "Сумма доходов за период"
            expense:
              type: string
              pattern: '^\d+\.\d{2}$'
              description: "Сумма расходов за период"
            balance:
              type: string
              pattern: '^\d+\.\d{2}$'
              description: "Итоговый баланс"
        currency:
          type: string
          pattern: '^[A-Z]{3}$'
          description: "Код валюты ISO-4217"
          default: "RUB"
        recentTransactions:
          type: array
          items:
            $ref: '#/components/schemas/TransactionDto'
    CategoryAnalyticsResponse:
      type: object
      properties:
        categories:
          type: array
          items:
            type: object
            properties:
              categoryId:
                type: string
                format: uuid
              categoryName:
                type: string
              amount:
                type: string
                pattern: '^\d+\.\d{2}$'
              percentage:
                type: number
                format: double
              transactionCount:
                type: integer
        totalAmount:
          type: string
          pattern: '^\d+\.\d{2}$'
        currency:
          type: string
          pattern: '^[A-Z]{3}$'
          default: "RUB"
    PeriodSummaryResponse:
      type: object
      properties:
        periods:
          type: array
          items:
            type: object
            properties:
              period:
                type: string
                format: date
                description: "Месяц в формате YYYY-MM"
              income:
                type: string
                pattern: '^\d+\.\d{2}$'
              expense:
                type: string
                pattern: '^\d+\.\d{2}$'
              balance:
                type: string
                pattern: '^\d+\.\d{2}$'
        currency:
          type: string
          pattern: '^[A-Z]{3}$'
          default: "RUB"
    DailyTransactionsResponse:
      type: object
      properties:
        days:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
                format: date
              income:
                type: string
                pattern: '^\d+\.\d{2}$'
              expense:
                type: string
                pattern: '^\d+\.\d{2}$'
              transactionCount:
                type: integer
        currency:
          type: string
          pattern: '^[A-Z]{3}$'
          default: "RUB"

    # Import/Export DTOs
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - BearerAuth: []